+++ changelog.md
@@ -5,6 +5,7 @@ Changelog
 -----
 
 - #35 added DEV_MODE. Set to mockup for easier development
+- #36 added RueckgratConfig for certralized config handling
 
 0.2.0
 -----